pyproject.toml
[project]
name = "kt-ai"
version = "0.1.0"
description = "KT.ai - AI-powered Project Onboarding & Knowledge Transfer platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "sqlalchemy>=2.0.0",
  "pydantic>=2.6.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "openai>=1.40.0",
  "numpy>=1.26.0",
  "faiss-cpu>=1.8.0",
  "hnswlib>=0.8.0",
  "llama-cpp-python>=0.3.0",
  "sentence-transformers>=2.7.0",
  "jira>=3.6.0",
  "python-docx>=1.1.2",
  "python-pptx>=0.6.23",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
